Leverage yup for ReportOptions validation
