+++ services/app-api/types/reports.ts
@@ -27,32 +27,6 @@ export interface ReportOptions {
   };
 }
 
-export const isReportOptions = (
-  obj: object | undefined
-): obj is ReportOptions => {
-  if (!obj) return false;
-  if (!("name" in obj) || "string" !== typeof obj.name) return false;
-  if (!("year" in obj) || "number" !== typeof obj.year) return false;
-  if (!("options" in obj) || "object" !== typeof obj.options) return false;
-  const extraKeys = Object.keys(obj).filter(
-    (key) => !["name", "year", "options"].includes(key)
-  );
-  if (extraKeys.length > 0) return false;
-
-  const options = obj.options;
-  if (!options) return false;
-  if ("cahps" in options && "boolean" !== typeof options.cahps) return false;
-  if ("hciidd" in options && "boolean" !== typeof options.hciidd) return false;
-  if ("nciad" in options && "boolean" !== typeof options.nciad) return false;
-  if ("pom" in options && "boolean" !== typeof options.pom) return false;
-  const extraOptions = Object.keys(options).filter(
-    (key) => !["cahps", "hciidd", "nciad", "pom"].includes(key)
-  );
-  if (extraOptions.length > 0) return false;
-
-  return true;
-};
-
 export interface CMIT {
   cmit: number;
   name: string;
